Revert "src: Refactor user space PM commands."
The MPTCP path management generic netlink API commands between the
upstream and multipath-tcp.org kernels will not be source compatible.
Move the refactored command implementation back
netlink_pm_mptcp_org.c.

This reverts commit dc92d3bb89e43695b21be80b28677cfcdbb41780.
